chore: move translations fetching logic into Header
